initial promotion building inside repository done
fixed a bug with product name parsing

+++ Repository/Interfaces/IPromotion.cs
@@ -9,6 +9,6 @@ namespace Repository.Interfaces
 
         int QuantityOffered {get;}
 
-        int PromotionalPricing {get;}
+        decimal PromotionalPricing {get;}
     }
 }
